Add an optional provision password that can be defined in the system -> variables. Create a new category called 'Provision' and add a variable to that category called 'Password'. If supplied then the url will require http://example.com/mod/provision/?mac=$MA&password=xxxxx
Also in the variables section I created a way to dynamically add variables that will be replaced in the template. The variables would be defined in the new 'Provision' category and must be prefixed with 'v_'. If the variable was v_example then its value would replace the {v_example} variable in the template.